+++ raspberryio/settings/base.py
@@ -346,3 +346,7 @@ except ImportError:
     pass
 else:
     set_dynamic_settings(globals())
+
+# Disqus
+DISQUS_SHORTNAME = 'raspberryio-dev'
+DISQUS_HOSTNAME = 'http://raspberry.io'