+++ docs/CHANGELOG.rst
@@ -1,6 +1,16 @@
 Raspberry IO Changelog
 ======================
 
+Version 0.3 (2013-09-05)
+------------------------
+
+* Improve coverage to near 100% (excepting aggregator)
+* Add DISQUS comments to project detail page
+* Add link to RaspberryIO Twitter account
+* Improve pagination throughout site
+* Fabfile fixes
+
+
 Version 0.2 - Open Source Release (2013-08-06)
 ----------------------------------------------